Handwriting Recognition
This app can simply reads a handwritten documents and determines the letters and numbers along with the space between words with the help of the model that we created. After determination you can see the result on the interface and also there is an option to save the result in a separate document.